backend/requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy[asyncio]
asyncpg
greenlet
psycopg2-binary
pydantic-settings
pillow
imagehash
opencv-python-headless
python-multipart
google-generativeai
boto3
celery
redis